Moved main.cpp inside folder and added working input capture
This adds functionality to capture key inputs from the terminal without having to use the enter key and will take single key inputs. The main function will constantly run, so I built in a stop system. All you need to do to stop the program is to hit key 97 (or 'a')